requirements.txt
numpy==2.2.6
pillow==12.1.0
